Create UserProfile component and apply logout actions
